Add levels

+++ src/main/java/frogger/model/Obstacle.java
@@ -3,7 +3,7 @@ package frogger.model;
 import javafx.scene.image.Image;
 
 public class Obstacle extends Actor {
-	private int speed;
+	private double speed;
 	@Override
 	public void act(long now) {
 		move(speed , 0);
@@ -13,7 +13,7 @@ public class Obstacle extends Actor {
 			setX(600);
 	}
 	
-	public Obstacle(String imageLink, int xpos, int ypos, int s, int w, int h) {
+	public Obstacle(String imageLink, int xpos, int ypos, double s, int w, int h) {
 		setImage(new Image(imageLink, w,h, true, true));
 		setX(xpos);
 		setY(ypos);
